Moved MyChartAPI to the lib directory.
