+++ lib/MyChartAPI.js
